New files to separate pipelines tests
